Refactor controllers and routes: fix bugs and remove unnecessary code

+++ src/controllers/tweet.controller.js
@@ -30,13 +30,15 @@ const updateTweet = asyncHandler( async (req, res) => {
         { new: true }
     );
 
-    if(!updatedTweet) throw new ApiError(500, "Failed to edit tweet. please try again!!!");
+    if(!updatedTweet) throw new ApiError(404, "Tweet not found");
 
     res.status(200).json( new ApiResponse(200, updatedTweet, "Tweet updated successfully"));
 });
 
 const deleteTweet = asyncHandler( async (req, res) => {
     const { tweetId } = req.params;
+
+    if (!mongoose.Types.ObjectId.isValid(tweetId)) throw new ApiError(400, "Invalid tweetId");
 
     await Tweet.findByIdAndDelete(tweetId);
 
@@ -48,6 +50,7 @@ const deleteTweet = asyncHandler( async (req, res) => {
 const getUserTweets = asyncHandler( async (req, res) => {
     const { userId } = req.params;
     const user = await User.findById(req.user.id).select('username avatarImage');
+    if (!user) throw new ApiError(404, "User not found");
 
     const tweets = await Tweet.aggregate([
         {
